pyproject.toml
[tool.poetry]
name = "tor_archivist"
version = "0.0.0"
description = "The officially licensed archivist for /r/TranscribersOfReddit"
authors = ["Grafeas Group <devs@grafeas.org>"]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9"
]

repository = "https://github.com/GrafeasGroup/tor_archivist"
homepage = "https://github.com/GrafeasGroup/tor_archivist"

[tool.poetry.dependencies]
python = "^3.9"
praw = "^7.6.0"
sh = "^1.14.3"
bugsnag = "^4.2.1"
requests = "^2.28.1"
slackclient = "^1"
python-dotenv = "^0.20.0"
blossom-wrapper = { git = "https://github.com/GrafeasGroup/blossom-wrapper.git", branch = "master" }
click = "^8.1.3"
shiv = "^1.0.1"

[tool.poetry.dev-dependencies]
better-exceptions = "^0.2.2"
pytest-cov = "^3.0.0"
flake8 = "^5.0.4"
mypy = "^0.971"
poetry2setup = "^1.0.0"

[tool.poetry.plugins."console_scripts"]
"tor_archivist" = "tor_archivist.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

